+++ iscc_sdk/__init__.py
@@ -2,7 +2,7 @@
 import os
 from platformdirs import PlatformDirs
 
-__version__ = "0.4.6"
+__version__ = "0.4.7"
 APP_NAME = "iscc-sdk"
 APP_AUTHOR = "iscc"
 dirs = PlatformDirs(appname=APP_NAME, appauthor=APP_AUTHOR)
@@ -24,3 +24,4 @@ from iscc_sdk.video import *
 from iscc_sdk.mp7 import *
 from iscc_sdk.exceptions import *
 from iscc_sdk.metadata import *
+from iscc_sdk.pdf import *